+++ Module.php
@@ -110,6 +110,26 @@ class Module extends \humhub\components\Module
     {
         return $this->settings->get('verifyPeerOff');
     }
+    public function getChat()
+    {
+        return $this->settings->get('chat');
+    }
+    public function getCompactHeader()
+    {
+        return $this->settings->get('compactHeader');
+    }
+    public function getFeedback()
+    {
+        return $this->settings->get('feedback');
+    }
+    public function getHelp()
+    {
+        return $this->settings->get('help');
+    }
+    public function getCompactToolbar()
+    {
+        return $this->settings->get('compactToolbar');
+    }
 
     /**
      *